fix order of stage.
